# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CodeFabric Admin API",
        "description": "CodeFabric 管理后台后端接口文档（Laravel 13）",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "本地开发环境"
        }
    ],
    "paths": {
...
        "/api/admin/roles": {
            "get": {
                "tags": [
                    "角色管理"
                ],
                "description": "获取角色列表",
                "operationId": "adminRoleList",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleListEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "角色管理"
                ],
                "description": "新增角色",
                "operationId": "adminRoleCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleItemEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "角色编码已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/roles/{id}": {
            "put": {
                "tags": [
                    "角色管理"
                ],
                "description": "编辑角色",
                "operationId": "adminRoleUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "编辑成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleItemEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "角色编码已存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "角色管理"
                ],
                "description": "删除角色",
                "operationId": "adminRoleDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/roles/{id}/permissions": {
            "post": {
                "tags": [
                    "角色管理"
                ],
                "description": "为角色分配菜单权限",
                "operationId": "adminRoleAssignPermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleAssignPermissionsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleItem": {
                "required": [
                    "id",
                    "name",
                    "code",
                    "userCount",
                    "permissionCount",
                    "menuIds",
                    "createdAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "超级管理员"
                    },
                    "code": {
                        "type": "string",
                        "example": "SUPER_ADMIN"
                    },
                    "description": {
                        "type": "string",
                        "example": "系统内置角色",
                        "nullable": true
                    },
                    "userCount": {
                        "type": "integer",
                        "example": 3
                    },
                    "permissionCount": {
                        "type": "integer",
                        "example": 18
                    },
                    "menuIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2026-04-05 10:00:00"
                    }
                },
                "type": "object"
            },
            "RoleListEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleItem"
                        }
                    }
                },
                "type": "object"
            },
            "RoleItemEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/RoleItem"
                    }
                },
                "type": "object"
            },
            "RoleCreateRequest": {
                "required": [
                    "name",
                    "code"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "运营",
                        "maxLength": 50
                    },
                    "code": {
                        "type": "string",
                        "example": "OPERATOR",
                        "maxLength": 50
                    },
                    "description": {
                        "type": "string",
                        "example": "运营角色",
                        "nullable": true,
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "RoleUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "运营",
                        "maxLength": 50
                    },
                    "code": {
                        "type": "string",
                        "example": "OPERATOR",
                        "maxLength": 50
                    },
                    "description": {
                        "type": "string",
                        "example": "运营角色",
                        "nullable": true,
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "RoleAssignPermissionsRequest": {
                "required": [
                    "menuIds"
                ],
                "properties": {
                    "menuIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "ApiEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "description": "CodeFabric 后端 OpenAPI 根定义。\n\n说明：\n- 使用 swagger-php Attributes 维护接口文档。\n- 业务返回统一为 { code, message, data }，通过组件 Schema 复用。",
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "type": "object",
                                "additionalProperties": true
                            },
                            {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "integer"
                            },
                            {
                                "type": "boolean"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "ApiErrorEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 2004
                    },
                    "message": {
                        "type": "string",
                        "example": "请求失败"
                    },
                    "data": {
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "在 Authorization 请求头中传入 Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "浏览器 Cookie 鉴权",
                "name": "admin_token",
                "in": "cookie"
            }
        }
    },
...
}